+++ Infrastructure.DataAccess/Infrastructure.DataAccess.csproj
@@ -1055,6 +1055,7 @@
     <Compile Include="Services\PocoTypeFromProxyResolver.cs" />
     <Compile Include="Services\WithinAmbienTransaction.cs" />
     <Compile Include="Services\TransactionManager.cs" />
+    <Compile Include="Tools\ConnectionStringTools.cs" />
     <Compile Include="Tools\SqlMigrationScriptRepository.cs" />
     <Compile Include="TypeConfigurationExtensions.cs" />
     <Compile Include="TypeMapping.cs" />
