+++ Infrastructure.DataAccess/KitosContext.cs
@@ -22,26 +22,14 @@ using Core.DomainModel.SSO;
 using Core.DomainModel.Notification;
 using Core.DomainModel.Tracking;
 using Core.DomainModel.UIConfiguration;
+using Infrastructure.DataAccess.Tools;
 
 namespace Infrastructure.DataAccess
 {
     public class KitosContext : DbContext
     {
-        public KitosContext() : this(GetConnectionString()) { }
-
-        //Create a method that uses ConfigurationManager to get the connection string from the web.config file, checks if the variable is base64 encoded, and decodes it if it is encoded
-        private static string GetConnectionString()
-        {
-            var connectionString = System.Configuration.ConfigurationManager.ConnectionStrings["KitosContext"]?.ConnectionString ?? "KitosContext";
-            if (!connectionString.StartsWith("base64:")) 
-                return connectionString;
-            
-            var base64EncodedString = connectionString.Substring("base64:".Length);
-            var base64EncodedBytes = Convert.FromBase64String(base64EncodedString);
-            connectionString = System.Text.Encoding.UTF8.GetString(base64EncodedBytes);
-            return connectionString;
-        }
-
+        public KitosContext() : this(ConnectionStringTools.GetConnectionString("KitosContext")) { }
+        
         public KitosContext(string nameOrConnectionString)
             : base(nameOrConnectionString)
         {
