+++ Presentation.Web/Startup.cs
@@ -14,6 +14,7 @@ using Presentation.Web.Infrastructure.Model.Authentication;
 using Presentation.Web.Ninject;
 using Presentation.Web.Infrastructure.Filters;
 using Presentation.Web.Infrastructure;
+using Infrastructure.DataAccess.Tools;
 
 [assembly: OwinStartup(typeof(Presentation.Web.Startup))]
 namespace Presentation.Web
@@ -51,26 +52,13 @@ namespace Presentation.Web
             app.Use<DenyTooLargeQueriesMiddleware>();
         }
 
-        //Create a method that gets kitos_HangfireDB connection string from the web.config file, checks if the variable is base64 encoded, and decodes it if it is encoded
-        private static string GetHangfireConnectionString()
-        {
-            var connectionString = System.Configuration.ConfigurationManager.ConnectionStrings["kitos_HangfireDB"]?.ConnectionString ?? "kitos_HangfireDB";
-            if (!connectionString.StartsWith("base64:")) 
-                return connectionString;
-            
-            var base64EncodedString = connectionString.Substring("base64:".Length);
-            var base64EncodedBytes = Convert.FromBase64String(base64EncodedString);
-            connectionString = System.Text.Encoding.UTF8.GetString(base64EncodedBytes);
-            return connectionString;
-        }
-
         private static void InitializeHangfire(IAppBuilder app)
         {
             // Initializing the Hangfire scheduler
             var standardKernel = new KernelBuilder().ForHangFire().Build();
 
             GlobalConfiguration.Configuration.UseNinjectActivator(standardKernel);
-            GlobalConfiguration.Configuration.UseSqlServerStorage(GetHangfireConnectionString());
+            GlobalConfiguration.Configuration.UseSqlServerStorage(ConnectionStringTools.GetConnectionString("kitos_HangfireDB"));
             GlobalJobFilters.Filters.Add(new AdvisSendFailureFilter(standardKernel));
             GlobalJobFilters.Filters.Add(new AutomaticRetryAttribute { Attempts = KitosConstants.MaxHangfireRetries });
